rename secretflow_fl to sfl, fix tests.fl
