+++ examples/security/v_tpa/custom_base/custom_sl_model.py
@@ -287,7 +287,7 @@ class CustomSLModel:
             verbose: 0, 1. Verbosity mode
             callbacks: List of Callback or Dict[device, Callback]. Callback can be:
             - `keras.callbacks.Callback` for tensorflow backend
-            - `from secretflow_fl.ml.nn.callbacks.callback.Callback` for torch backend
+            - `from sfl.ml.nn.callbacks.callback.Callback` for torch backend
             validation_data: Data on which to validate
             shuffle: Whether shuffle dataset or not
             validation_freq: specifies how many training epochs to run before a new validation run is performed
