+++ pyproject.toml
@@ -19,6 +19,7 @@ requires-python = ">=3.10,<3.12"
 dynamic = ["version"]
 
 dependencies = [
+    "polars",
     "numpy",
     "pandas==1.5.3",
     "scikit-learn==1.5.2",
@@ -57,6 +58,8 @@ dev = [
     "psutil",
     # added for FreqFed
     "hdbscan",
+    "matplotlib",
+    "tabulate",
 ]
 
 [project.urls]
