+++ sfl/component/ml/nn/sl/base.py
@@ -158,7 +158,7 @@ DEFAULT_MODELS_CODE = """\
 # from tensorflow import Module, keras
 # from tensorflow.keras import Model, layers
 # from tensorflow.keras.layers import Layer
-# from secretflow_fl.ml.nn import applications as apps
+# from sfl.ml.nn import applications as apps
 
 def create_base_model(input_dim, output_dim):
     model = keras.Sequential(
